Correct SOS service pack/publish behavior
The SOS service is calling SOS_publish(...) and transmitting data to the daemon (apparently) as independent actions for each element of each snapshot. The desired behaviour is for calls to SOS_pack_related(...) to enqueue these values in the (SOS_pub*) sos_pub handle, and only once at the triggering event connect and transmit all enqueued values.